requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
langchain==0.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
numpy==1.24.3